// src/app/modules/crm-mobile/Interfaces/year.interfaces.ts
export interface Year {
    YearMasterSid: number;
    CompanyMasterSid: number;
    YearName: string;
    YearCode: number;
    StartDate: any;
    EndDate: any;
    CurrentYear: string;
    YearEndCompleted: string;
    status: string;
    Remarks: string;
    createdOn: string;
    updatedOn: string;
    createdBy: string;
    updatedBy: string
}